Refactor health configuration.
